Make System::FindStaffIndex handle the case where the staff is not found. Update ScoreArea::drawStdNotation to use FindStaffIndex
git-svn-id: http://svn.powertabs.org/svn/ptedev/trunk@255 5b1460fe-822e-4aa1-ac1b-8a6412c92c8c

+++ source/powertabdocument/system.h
@@ -228,7 +228,7 @@ public:
     Staff* GetStaff(uint32_t index) const
     {CHECK_THAT(IsValidStaffIndex(index), NULL); return (m_staffArray[index]);}
 
-    int FindStaffIndex(Staff* staff) const;
+    size_t FindStaffIndex(Staff* staff) const;
 
     // Barline Functions
     /// Determines if a barline index is valid
